implemented first version of unit formation movement, NEED REWORK,
//currently they all get parented to a new object that gets its own movement comp and moves all the units.
The problem is that this looks really clunky and wrong since now verey ship si pivoting on one median point instead of their indiviual points.
A potential alternative would be to have the commander hold the group formation object with each objects goal beacon.
The other ships would then inherit the speed from the commander ship and try to reach their goal beacon, which is attached to the commander.
The other ships would only count as havign reached their target when the commander ship has reached its target and they are ontop of theiw goal transforms.